Moved outliner tool
