+++ Arma3ObjectBuilder/tool_outliner/outliner.py
