+++ Arma3ObjectBuilder/ui/__init__.py
@@ -8,7 +8,6 @@ from . import import_export_paa
 from . import props_action
 from . import props_material
 from . import props_object_mesh
-from . import tool_outliner
 from . import tool_mass
 from . import tool_proxies
 from . import tool_rigging
