+++ Arma3ObjectBuilder/utilities/__init__.py
@@ -1,6 +1,5 @@
 # In order of dependency
 from . import proxy
 from . import masses
-from . import outliner
 from . import rigging
 from . import structure
